starting refactor 404 handler
